Written the core of the driver for testing purposes
The driver in this state initializes everything and simply passes the
values from the physical joystick to the virtual one. The next steps are
to make 2 joysticks appear as one virtual, add remapping to the common
layout, and add a script that will find the FJoys, remove their event
handlers, and pass their adresses to the driver.